Re-organises ram-related module
